+++ nmanga/cmd.py
@@ -35,7 +35,7 @@ from .cli.image_optimizer import image_optimizer
 from .cli.image_tagging import image_tagging
 from .cli.manual_split import manual_split
 from .cli.merge_chapters import merge_chapters
-from .cli.releases import prepare_releases
+from .cli.releases import prepare_releases, prepare_releases_chapter
 from .cli.spreads_manager import spreads
 from .constants import __author__, __name__, __version__
 from .term import get_console
@@ -85,6 +85,7 @@ main.add_command(pack_releases)
 main.add_command(pack_releases_epub_mode)
 main.add_command(pack_releases_comment_archive)
 main.add_command(prepare_releases)
+main.add_command(prepare_releases_chapter)
 main.add_command(spreads)
 main.add_command(image_tagging)
 main.add_command(image_optimizer)